refactor: create download file function
